+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.2.0 (unreleased)
 
 - Changed pattern for fitting models - use `Model::params()` instead of `Model::new()`
-- Changed `fit`, `fit_eval`, `save`, and `load` to return `Result`
+- Changed `fit`, `fit_eval`, `cv`, `save`, and `load` to return `Result`
+- Changed `cv` to return average error
 
 ## 0.1.1 (2021-07-27)